add: exception handling in the delete post

+++ app/graphql/mutations/post/delete_post.rb
@@ -7,9 +7,14 @@ module Mutations
       type Types::PostType, null: false
 
       def resolve(id:)
-        post = ::Post.find(id)
-        post.destroy!
-        post
+        post = ::Post.find_by(id: id)
+
+       if post # truthy value
+          post.destroy!
+          post
+       else
+          raise ::NotFoundError, "Post with id #{id} not found"
+       end
       end
     end
   end
